pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.1",
]
